Refactored test cases and added helper functions following addition of sqs payload types
