+++ integration-tests/utils/aws/invokeSQSOperationsLambdaFunction.ts
@@ -1,12 +1,13 @@
 import { getIntegrationTestEnvironmentVariable } from '../getIntegrationTestEnvironmentVariable'
 import { InvokeCommand, LambdaClient } from '@aws-sdk/client-lambda'
+import {
+  CreateResultFileSQSPayload,
+  QueryCompleteSQSPayload
+} from '../types/sqsPayload'
 
-export const invokeSQSOperationsLambda = async (payload: {
-  athenaQueryId: string
-  recipientEmail: string
-  recipientName: string
-  zendeskTicketId: string
-}) => {
+export const invokeSQSOperationsLambda = async (
+  payload: CreateResultFileSQSPayload | QueryCompleteSQSPayload
+) => {
   const input = {
     FunctionName: 'txma-qr-dev-tools-sqs-operations',
     Payload: jsonToUint8Array(payload)
@@ -18,18 +19,14 @@ export const invokeSQSOperationsLambda = async (payload: {
   await lambdaClient.send(new InvokeCommand(input))
 }
 
-const jsonToUint8Array = (json: {
-  athenaQueryId: string
-  recipientEmail: string
-  recipientName: string
-  zendeskTicketId: string
-}): Uint8Array => {
+const jsonToUint8Array = (
+  json: CreateResultFileSQSPayload | QueryCompleteSQSPayload
+): Uint8Array => {
   const string = JSON.stringify(json, null, 0)
   const uint8Array = new Uint8Array()
 
   for (let i = 0; i < string.length; i++) {
     uint8Array[i] = string.charCodeAt(i)
   }
-
   return uint8Array
 }
