+++ integration-tests/utils/types/integrationTestEnvironmentVariable.ts
@@ -9,4 +9,5 @@ export type IntegrationTestEnvironmentVariable = {
     | 'QUERY_COMPLETED_QUEUE_URL'
     | 'RECIPIENT_EMAIL'
     | 'RECIPIENT_NAME'
+    | 'INTEGRATION_TESTS_TRIGGER_QUEUE_URL'
 }
